Fix decode only with parities
When decoding only with parities (e.g. k=10, m=10 and decode with
10 parity fragments and no data fragments), current liberasurecode
will give up to decode with a failure because liberasurecode retrieves
original content length and payload size only from a "DATA" fragment.

This patch allows liberasurecode to retrieve also a parity fragment
(if the available fragments consist of only parity fragments) and then
fixes the bug.

Note that this failure happens only when the number of parity fragments
equals to or more than the number of data fragments. (i.e. m >= k)